:recycle: refactoring: 스타일 분리
