+++ client/src/components/Layout/LeftNav/LeftNav.tsx
@@ -1,7 +1,7 @@
-import styled from 'styled-components';
 import { Link } from 'react-router-dom';
 import { Logo } from '../../../assets/Icons';
 import HamburgerMenu from './HamburgerMenu';
+import { Container, Nav } from './style';
 
 function LeftNav() {
 	return (
@@ -15,18 +15,5 @@ function LeftNav() {
 		</Container>
 	);
 }
-
-const Container = styled.div`
-	position: relative;
-`;
-
-const Nav = styled.nav`
-	position: sticky;
-	top: 45px;
-	margin: 45px 0 0 40px;
-	display: flex;
-	flex-direction: column;
-	width: 100px;
-`;
 
 export default LeftNav;
