+++ client/src/components/Layout/LeftNav/MenuCategory.tsx
@@ -1,7 +1,7 @@
-import styled, { keyframes } from 'styled-components';
 import { Link } from 'react-router-dom';
 import { CATEGORIES } from '../../../assets/Constants';
 import LeftNavMenu from './LeftNavMenu';
+import { CategoryContainer } from './style';
 
 export default function MenuCategory() {
 	return (
@@ -19,21 +19,3 @@ export default function MenuCategory() {
 		</CategoryContainer>
 	);
 }
-
-const openList = keyframes`
-	0% {
-		opacity: 0;
-		top: -20px;
-	}
-	100% {
-		top: 0;
-		opacity: 100%;
-	}
-`;
-
-const CategoryContainer = styled.ul`
-	display: flex;
-	flex-direction: column;
-	position: relative;
-	animation: ${openList} 0.3s ease-in-out;
-`;
